deposit: removes validator view
 * Closes #902.
