--- cap/modules/deposit/views/__init__.py
@@ -1,1 +1,0 @@
-"""Deposit views."""
